api/requirements.txt
flask>=3.0
flask-cors>=4.0
anthropic>=0.40
pandas>=2.0
openpyxl>=3.1
python-docx>=1.0
pdfplumber>=0.10
PyMuPDF>=1.24
python-dotenv>=1.0
supabase>=2.0
